+++ Runtime/Object Pool/PoolableParticleSystem.cs
@@ -4,6 +4,9 @@ namespace LMirman.Utilities
 {
 	/// <summary>
 	/// Built in implementation for recycling particle systems using the <see cref="ObjectPool"/> system.
+	/// <br/>
+	/// Will automatically set the particle system's stop action to callback and return it to the pool when the stop callback is invoked.
+	/// When the object is retrieved from the pool it will automatically play the system again and (optionally) clear old particles from the previous instance.
 	/// </summary>
 	[RequireComponent(typeof(ParticleSystem), typeof(Poolable))]
 	public class PoolableParticleSystem : MonoBehaviour